+++ rust/cubesql/cubesql/src/compile/rewrite/mod.rs
@@ -95,6 +95,11 @@ crate::plan_to_language! {
             n: usize,
             input: Arc<LogicalPlan>,
         },
+        TableUDFs {
+            expr: Vec<Expr>,
+            input: Arc<LogicalPlan>,
+            schema: DFSchemaRef,
+        },
         CreateExternalTable {
             schema: DFSchemaRef,
             name: String,
@@ -178,12 +183,20 @@ crate::plan_to_language! {
             fun: Arc<AggregateUDF>,
             args: Vec<Expr>,
         },
+        TableUDFExpr {
+            fun: Arc<TableUDF>,
+            args: Vec<Expr>,
+        },
         InListExpr {
             expr: Box<Expr>,
             list: Vec<Expr>,
             negated: bool,
         },
         WildcardExpr {},
+        GetIndexedFieldExpr {
+            expr: Box<Expr>,
+            key: ScalarValue,
+        },
 
         CubeScan {
             cube: Arc<LogicalPlan>,
@@ -253,6 +266,12 @@ crate::plan_to_language! {
             members: Vec<LogicalPlan>,
             aliases: Vec<(String, String)>,
             table_name: Option<String>,
+        },
+        InnerProjectionSplitReplacer {
+            members: Vec<LogicalPlan>,
+        },
+        OuterProjectionSplitReplacer {
+            members: Vec<LogicalPlan>,
         },
     }
 }
@@ -506,6 +525,10 @@ fn column_expr(column: impl Display) -> String {
     format!("(ColumnExpr {})", column)
 }
 
+fn cast_expr(expr: impl Display, data_type: impl Display) -> String {
+    format!("(CastExpr {} {})", expr, data_type)
+}
+
 fn alias_expr(column: impl Display, alias: impl Display) -> String {
     format!("(AliasExpr {} {})", column, alias)
 }
@@ -555,6 +578,14 @@ fn order_replacer(members: impl Display, aliases: impl Display, cube: impl Displ
 
 fn filter_replacer(members: impl Display, cube: impl Display) -> String {
     format!("(FilterReplacer {} {})", members, cube)
+}
+
+fn inner_projection_split_replacer(members: impl Display) -> String {
+    format!("(InnerProjectionSplitReplacer {})", members)
+}
+
+fn outer_projection_split_replacer(members: impl Display) -> String {
+    format!("(OuterProjectionSplitReplacer {})", members)
 }
 
 fn cube_scan_members(left: impl Display, right: impl Display) -> String {